+++ dash/dash-renderer/src/types/component.ts
@@ -1,12 +1,39 @@
+import {
+    DashContext,
+    DashContextProviderProps,
+    useDashContext
+} from '../wrapper/DashContext';
+import ExternalWrapper from '../wrapper/ExternalWrapper';
+import {stringifyId} from '../actions/dependencies';
+import {
+    DevtoolContext,
+    useDevtool,
+    useDevtoolMenuButtonClassName
+} from '../components/error/menu/DevtoolContext';
+
 export type BaseDashProps = {
     id?: string;
+    componentPath?: DashLayoutPath;
     [key: string]: any;
 };
+
+export interface DashComponentApi {
+    ExternalWrapper: typeof ExternalWrapper;
+    DashContext: typeof DashContext;
+    useDashContext: typeof useDashContext;
+    getLayout: (componentPathOrId: DashLayoutPath | string) => any;
+    stringifyId: typeof stringifyId;
+    devtool: {
+        DevtoolContext: typeof DevtoolContext;
+        useDevtool: typeof useDevtool;
+        useDevtoolMenuButtonClassName: typeof useDevtoolMenuButtonClassName;
+    };
+}
 
 export type DashComponent = {
     type: string;
     namespace: string;
-    props: BaseDashProps;
+    props: DashContextProviderProps & BaseDashProps;
 };
 
 export type UpdatePropsPayload = {